fix: update handlers to refactored pricefeed
